contact-form/requirements.txt
uvicorn
fastapi
sendgrid
dotenv
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi